update for all the display type
add the support to all display type:
S011, S014 (still not test), S021, S031;
Lectum, Tricolor and Legio
combine UC8156 and UC8179 together

+++ utils.c
@@ -154,7 +154,7 @@ int parser_read_area(const char *str, const char *sep, struct area *a)
 
 int parser_read_file_line(FIL *f, char *buffer, int max_length)
 {
-	size_t count;
+	uint16_t count;
 	char *out;
 	int i;
